+++ src/components/Textarea/Textarea.test.js
@@ -1,4 +1,4 @@
-import { shallow } from "enzyme";
+import { mount, shallow } from "enzyme";
 import React from "react";
 
 import Textarea from "./Textarea";
@@ -14,5 +14,10 @@ describe("Textarea ", () => {
     const className = wrapper.find("textarea").prop("className");
     expect(className.includes("p-form-validation__input")).toBe(true);
     expect(className.includes("extra-class")).toBe(true);
+  });
+
+  it("can take focus on first render", () => {
+    const wrapper = mount(<Textarea takeFocus />);
+    expect(wrapper.find("textarea").getDOMNode()).toBe(document.activeElement);
   });
 });