+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
+### Added
+
+- Updated Input, Select and Textarea to have a takeFocus prop which sets the focus to the input on first render.
+
 ## [0.8.1] - 2020-07-16
 
 ### Changed